requirements.txt
streamlit
requests
pathway
google-generativeai
pandas
pydantic